requirements.txt
pytorch-lightning==2.4.0
torch>=2.2.0
torchvision>=0.17.0
tensorboard>=2.14
numpy
opencv-python
matplotlib
einops
pyyaml
pytest